requirement.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
distlib==0.3.1
Django==3.0.8
djangorestframework==3.11.0
filelock==3.0.12
get==0.0.0
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
post==0.0.0
psycopg2==2.8.5
pylint==2.5.3
pytz==2020.1
query-string==0.0.0
request==0.0.0
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==20.0.25
virtualenvwrapper-win==1.2.6
wrapt==1.12.1
zipp==3.1.0
